Refactor transition state class params
