+++ packages/core/index.js
@@ -3,15 +3,18 @@ export * from "./src/js/CollectionEntry";
 export * from "./src/js/Breakpoint";
 export * from "./src/js/FocusTrap";
 
-export * from "./src/js/cssVar";
+export * from "./src/js/utilities/toCamel";
+export * from "./src/js/utilities/toKebab";
+export * from "./src/js/utilities/toMilliseconds";
+
+export * from "./src/js/helpers/cssVar";
+export * from "./src/js/helpers/getPrefix";
+
 export * from "./src/js/getConfig";
 export * from "./src/js/getCustomProps";
 export * from "./src/js/getElement";
-export * from "./src/js/getPrefix";
 export * from "./src/js/localStore";
 export * from "./src/js/teleport";
 export * from "./src/js/themeStore";
-export * from "./src/js/toCamel";
-export * from "./src/js/toKebab";
 export * from "./src/js/transition";
 export * from "./src/js/updateGlobalState";
