+++ packages/core/src/js/utilities/toCamel.js
