+++ packages/core/src/js/utilities/toKebab.js
