+++ packages/drawer/src/js/close.js
@@ -15,13 +15,13 @@ export async function close(query, transitionOverride, focus = true) {
 
     // Run the close transition.
     if ((transitionOverride != undefined) ? transitionOverride : entry.getSetting("transition")) {
-      await transition(entry.el, {
-        start: entry.getSetting("stateOpening"),
-        finish: entry.getSetting("stateOpened")
-      }, {
-        start: entry.getSetting("stateClosing"),
-        finish: entry.getSetting("stateClosed")
-      }, entry.getSetting("transitionDuration"));
+      await transition(
+        entry.el, 
+        entry.getSetting("stateOpened"),
+        entry.getSetting("stateClosing"),
+        entry.getSetting("stateClosed"),
+        entry.getSetting("transitionDuration")
+      );
     } else {
       entry.el.classList.add(entry.getSetting("stateClosed"));
       entry.el.classList.remove(entry.getSetting("stateOpened"));
